The final version of the game engine.
The library developer, there are two types of modules: logic and
peripherals. In logic receives packets from the periphery and the
decision about which package and which module to send. Peripherals,
received a package from the logic of its flow is the function that is
described in the package of logic.